An experiment for how to write a LLVM pass in Coq. We take the simplest pass in LLVM/lib/Transforms/Scalar/ConstantProp.cpp which does a simple constant propagation.
1) The LLVM pass is written in C++, imperatively, without termination guards.
It follows Kildall's generic data flow analysis algorithm: a loop works on a worklist, until
the worklist is empty, which indicts a fixed point of the analysis.

To define this in Coq, we have at least three options:
  a) Like CompCert's, a step-bound is used to make Coq recursion structured. The extracted OCaml
     terminates when the bound meets. Because we know all of these known algorithms are finite,
     as long as we pick a large-enough bound, or the largest step number in the worst case, the
     extracted code wont have any 'false-positive' result. 

  b) Define the Coq recursion with well-founded condition or measure functions. This can be
     done by 'Function' or 'Program'. For ConstantProp, the measure function is (#instr)*(#instr)+(#worklist).
     I found Program is slightly smarter than Function. Program can provide better error msgs,
     more flexible support for the equality when pattern-matching, and simpler obligations to fix.
     The extracted OCaml code looks more like C++ code, because it does not have the step-bound
     after the automatic elimination of proofs.

     I am using this solution.

  c) "Fixed Point Semantics and Partial Recursion in Coq" proposed a 'domain-theory' Tarski-fixed point
     based solution. It seems more powerful. But 'Function' and 'Program' are sufficient for the time being.
     And the proposed idea is not easy to understand because of the 'domain-theory'...

We cannot program in Coq exactly same as in C++. At least, the while loop in C++ should be
replaced by a recursive OCaml function, because a recursive Coq function is extracted into a recursive
OCaml, unless we are designing a domain-specific extraction from Coq to C++...

2) How to represent 'imperative programming' in Coq.
The normal encoding of Monads is not quite suitable for reasoning about properties.
For example, if we notate "do s <- a ; b" as "(\s. b) a", when reasoning about the body b,
we may need to know the equality "a = s", but Coq does not give us this for free. 

The known trick to get an equality hypothesis is by 'pattern matching. We can 
use '(match a as a' return (a = a' -> _) with a'' => fun H:(a=a'') => b end) (equal_refl a).
In this encoding, we have the equalities for free.

And in the context, which 'Program' generates for us to prove 'obligations', those equalities
are listed sequentially, which provides the exact information to prove imperative properties.

3) What granularity of LLVM interfaces is exposed?
This granularity affects how much Coq code we are writing, and how many LLVM axioms we are trusting.
Suppose we take Coq definition as a specification of a pass, the trade off means how formal the 
specification will be.

Currently, I randomly selected the exposed LLVM interfaces. But the LLVM OCaml bindings have not been done yet.
This is the TODO.

git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2251 eef5ea85-cc2e-0410-bd84-0fae1f012da8