fixed some CoqLLVM bindings when designing OCamlLLVM bindings.
Problem 1 was how to represent stl::data_structures,
such as set, map, list, ...
  1) We use the native data structures in Coq
  2) We wrap them in OCaml bindings
I did the 2nd one, the first one may be better?
But their interfaces are not exactly matched,
the algorithm replies on C++ STL interfaces,
we should translate these interfaces if we do the 1st.

Problem 2,
some C++ functions must be rewritten to indict sequential relations,
  Worklist wl; wl.erase(I);
needs to be
  wl' = worklist_erase wl I
otherwise, we dont know the wl after 'erase' is different from 
the one before 'erase'. The rewritten function tracks the
states for free.

Problem 3:
There are too many match-return-end in the Coq,
it is great if we can define any notation to hind them.


git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2252 eef5ea85-cc2e-0410-bd84-0fae1f012da8