Example LLVM pass authored entirely in ocaml.  For now the pass prints to the console for each BasicBlock it encounters.  Most of the work here is getting C, C++, and OCaml all talking with each other and holding hands inside of LLVM.  Woot.
Currently there is a blocking issue with including libLLVMCore.a in the mix.  This is troublesome because we need Core to access the C and OCaml bindings for the LLVM IR.  I'm waiting for a response from the llvmdev mailing list before moving forward.

Alternatively, we can author bindings ourselves from C++ to OCaml.  Note we may have to do anyways because the OCaml bindings are simply extensions of the imperative C bindings.

(This directory should be placed in <llvm install directory>/lib/Transforms for building purposes.  You can execute the pass by using opt with an incantation similar to:

opt -load "llvm2.6/Debug/lib/OCamlPass.so" -ocamlpass test.bc)


git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2285 eef5ea85-cc2e-0410-bd84-0fae1f012da8