1) redefined Coq/OCaml bindings for proofs. If a function does an side-effect to an input, it should be taken as an output.
2) Problems for proofs. Although the correctness of the constant
propagation is obvious, this is not true in Coq. First, the definition
of constant propagation is not a simple structural recursion, 
it is well-founded because we can define a measure function, which
decreases during nested recursions. Second, we need to use 
dependent equalities. Finally, we need well-founded inductive
pricinples from the functions. 

'Function' has 1) and 3)(partially, its inductive scheme can do
for structural recursion, but not well-found measure..), 
but cannot do 2) if we use dependent equalities too much. Program 
Fixpoint is good for 1)2), but doesn't support 3).

Compert defined such recursion on steps with a max step num.
The recursion fails if the max num is reached, but it is not
failed because the recursion defined is terminated. Should we
do this too?

3) My OpenSUSE 11.2 requires -lreadline when compiling libLLVMPass.so,
which was not required on Eniac. I added it locally.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2353 eef5ea85-cc2e-0410-bd84-0fae1f012da8