1) added one missing file "llvm-2.6/include/llvm-c/OCamlPass/ConstantPropTv.h"
2) created vol/src/ssa/llvm_op, and moved relative files into llvm_op.
This llvm_op contains the static and dynamic semantics (small-steps) of
LLVM. The problem of the semantics is that they tried to do exactly the
same definitions as LLVM's implementations which are highly imperative,
and practical. Therefore, these definitions look clumsy from the view of
formalization, and they may not be suitful to formal proofs. From this
checkin, we are switching to new definitions with formal proofs in mind, 
llvm_op should be an application of the new definitions.

3) Makefiles were also changed to match 2).



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2456 eef5ea85-cc2e-0410-bd84-0fae1f012da8