1) removed 'ExitValue' from State, which was to denote the return value of 'main'. Actually this return result is already stored in the return insn on the top of the stack, because if a function stops, it must stop with the return insn as its last instruction (unless supporting 'invoke' to unwind stacks).
2) removed the cases that check if the recent stack is empty when return
is the current insn. These cases are too specific as inductive rules
(not strong enough for proofs). We can always check if there is only one
frame in the stack, and if the last insn on this frame is return to see
if we are reaching the final state.

3) Also commented out 'return void' to simply proofs, 'return typ value'
gives all interesting cases.

4) split transitivity-reflexivity closures into two cases: with 0, or
more steps -->*, and with at least one steps -->+. And generalized divergence
s -->* oo  as iff ex s', s.t. s -->+ s' /\ s -->* oo. (oo denotes
infinity, -->* oo is [dsop_diverges] in Coq), rather than ex s', s.t. s-->s' /\ s --->* oo. 

This change made coinductive proofs work at the following proof:
  if s ==> oo, then s -->* oo. where ==> is [dbopInf] in Coq.
s ==> oo gives s ==> s' and s' ==> oo.
By s ==> s' implies s -->+ s', we know s -->+ s'. With the new definion
of -->* oo, we conclude the proof by coinduction hypothesis, because
this CIH is used as one of arguments in this constructor, and satisfies
'Guarded' conditions. 

However, with the old definition, we have to complete the proof with
auxillary facts [trans]: s'' -->* s' /\ s' -->* oo -> s'' -->* oo, where
s --> s'' and s'' -->* s' by inversion on s -->+ s'. In this case,
by CIH, we have s' -->* oo. But s' -->* oo is first used in [trans], and
[trans] is used as one of the arguments of -->* oo. This fails 'Guarded'
checking in Coq, because it is hard to analyze the relations between CIH
results and constructors with [trans] in between.

5) Fixed the det-big-step cases "dbFdef" and "dbCallInsn", to match
small-steps, dbCallInsn calls dbFdef with the first insn in this
function on the top of the present stack ECs, as dbFdef' initial state,
and dbFdef stops when 'return' is the insn, and the returning stack is
ECs, finally the control returns back to dbCallInsn that gets the return
result, the new memory status by dbFdef, and then updates its memory
w.r.t the return result. This process exactly matches the sequence from
dsCallInsn to its matching dsReturn.

dbReturn was also removed, because we will never evalute 'return'
explicitly, that case is considered by 'dbCallInsn' when 'dbFdef' is
done.

also simplied rules.

7) the similar changes were done for non-deterministic rules, but with no
proofs and checkings.

8) Added tactics.v that defines cases notations for induction proofs

9) added opsem_conv.v that proved that bigsteps imply smallsteps when
convergence, going-wrong and divergence.




git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2473 eef5ea85-cc2e-0410-bd84-0fae1f012da8