1) simplified subblock-opsem: removed invariant variables from states; big-step opsem doesn't need to take the whole stack as a state, the current frame is sufficient. llvm-opsem can be simplified too.
2) defined TV for blocks and functions, proved correctness within a
block. 

We need one more property of SE: weakening or strengthening
symbolic mapping with unused variables (that are not used or defined by
SE) does not change results, and this should also be true for the denotation of
SE. The opsem of each function does not know how many local
variables are defined within this function, which updates or adds
defined variables at runtime. So when defining TV for functions, we
don't know which variables SE are using for each subblock (We need to
let SE match opsem that works on only used/defined variables). We can take
all variable in this function for inputs of SE, and then weaken or strenthen them at each
subblock to prove the its result denotes opsem. We can also extend the syntax and opsem to get the list of local
variables with undefined values at the beginning of a function, then SE
doesn't need to hold the above property.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2534 eef5ea85-cc2e-0410-bd84-0fae1f012da8