1) split the ssa.v generated from Ott into ssa_def.v/ssa_lib.v/ssa_static.v at ./ssa and ./ssa/llvm_op, to break circular dependency between extractable Ocaml and Ocaml auxulary modules. Now, Ocaml aux libs only need ssa_def, but ssa_lib can use these Ocaml aux libs. Updated lib/OCamlPass/CoCaml wrt this change.
2) connected memory operations between Llvmcaml.ml and
llvm_executionengine.ml. The type 'val' is only used in CompCert Mem,
and 'GenericValue' is only used in ssa_dynamic, such that we can clearly
extract 'GenericValue' into LLVM's GenericValue. We extracted 'mem' to
be ExecutionEngine.t * Llvm.llmodule that are sufficient to manage
memory.

added ssa_main.ml, which is the main entry of the interperter.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2691 eef5ea85-cc2e-0410-bd84-0fae1f012da8