supported external functions 1) in Coq, we assumed that there are axioms that define behaviors of external functions. FIXME: We should add each external call with its inputs and outputs as a trace event. 2) in LLVM, although interpreter does not support external calls, we extended Ocaml bindings by reusing external call support in JIT, which searches external calls from dynamic libraries.
Worked around a bug of LLVM. Some of its targets, say X86, cannot output
correct target datalayouts information in IRs, although internally the
datalayout are hard-coded in code. Trying to create TargetData objects
from values in IR voilates assertions in Module::setDatalayout. For
example X86 outputs i64:32:64, which is supposed to be i64:64:64. To
fix it, we extracted LLVMtd.TargetData to be ExecutionEngine.t that
contains the correct TD layout, rather than generating TD from the
buggy string in IR.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2697 eef5ea85-cc2e-0410-bd84-0fae1f012da8