supported function pointers 1) changed call insn to be      id = .. .. call value ..    from      id = .. .. call fid ..
  This allows function name to be any value that can potentially point
to an address where function is, namely a function pointer.

2) In Coq, we maintain a FunTable that maps function names to their
addresses that are taken as function pointers. When we are calling a
function via an id, we first search 'lookupFdefViaGVFromFunTable' in Globals via the value id to get its address, and then search in FunTable to get its name, via the name, we search in CurProducts to get its definition. 

   We assume that there is an 'initFunTable' that returns function addresses to initialize FunTable.

3) To realize 'lookupFdefViaGVFromFunTable' in LLVM, we can try to
dynamically cast fptr to Function*, if failed, return None; if
successeful, we can return this function's name.
  
'initFunTable' is realized to be 'ExecutionEngine.get_pointer_to_function'.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2715 eef5ea85-cc2e-0410-bd84-0fae1f012da8