1) fixed a bug when llvm pretty printing 'call' instructions 2) report.lli.txt is the test result of llvm interpreter, which also runs slow in most of the cases lynx runs slow
  Callgrind reports the costs of lynx, 57% is from mark_slice, 21% is
from sweep_slice. This means GC is not working well, because of some
wrong config. alloc_custom allows users to specify 'used' and 'max', the
radio used/max is the frequency of GC. We are using the default setting,
used=0/max=1, which means GC should run as slow as possible. Do not
understant why the cost is so high.

  5% of cost is from compare_val, which is called by caml_equal.

  Callgrind also reports the costs of lli, which does not have very
expensive function calls. The highest self costs are from STL libraries,
and 7% is used by STL's allocator. To design an ad-hoc, rather than
generic vector or map to replace STL's, should be able to reduce 10% of
the overhead.

3) Started to support recursive types. This version only has global
named types, not defined the unification, construction and datalayout
calculation yet.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2717 eef5ea85-cc2e-0410-bd84-0fae1f012da8