1) implemented the metadata inference algorithm that reads a Softbound ouput program, automatically guesses which variables are base/bound, and addresses of base/bound for __loadHashMetadata, and creates "metadata.db".
It is trivially terminating because it is defined by the numbr of computation 
steps.

The new feature of 8.3 extraction allows us to map Coq.nat (computation
steps) to OCaml.int efficiently.

TODO: Given the inferred metadata, we should validate that they are
"well-formed", which means the inference algorithm stopped because some
invariant holds, rather than out of computation steps. This
well-formedness is also key to prove the correctness of the TV.

2) Two new features of the TV
  a) Checked that size_of_type is correct at each dereference check, which
should be 
  i32 ptrtoint(typ* getelementptr(typ* null, i32 1) to i32)
if accessing a typ* pointer.

  b) Checked that if a function's argment is pointer, the pointer's
metadata passed in the function through additional arguments are valid.
We do not need to specify which additional arguments are for metatdata,
which can be inferred.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2821 eef5ea85-cc2e-0410-bd84-0fae1f012da8