checkpoint before some more updates of mem-simulation.
NOTE: found an inconsistent between this proof and the old proofs. The
old proofs assume that malloc zeros-out all metadata indiced by the
address in the allocation range. Otherwise, if malloc reuses free-ed
addresses that have old metadata, then runtime can load wrong metadata
for the reused address.

The SoftBound implementation for llvm-2.6 does not create wrappers for
malloc, so the above assumption (zero-out metadata) is not true. This is
because llvm 2.6 takes malloc as a primitive insn. SoftBound supports
malloc wrapper for llvm 2.7>=, which takes malloc as intrinsic
functions.

This is not a problem for the current memory model because it has
infinite memory, and does not reuse any deallocated address. But this
must be a problem to prove CETS for llvm 2.6 since we need a more
realistic memory model that reuses memory.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2881 eef5ea85-cc2e-0410-bd84-0fae1f012da8