redefined domination lattice by ListSet.
The lattice by FSet is good for the fsetdec tactics, but makes inductive
principles over FSet hard to use. The problem is FSet interface is not
affected by the order of its underlying elememts, so if we use its
'fold' to compute anything, the fold's ind-rule has to ensure this
invariant by showing that relative properties are instances of
'Equivalence'.

The ListSet impl fixed the problem, while we need to define new tactics
to automate proofs.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2950 eef5ea85-cc2e-0410-bd84-0fae1f012da8