fixed a bug in the llvm2coq translation, which should check global variables are always of pointer typ, so we can safely use its element typ. In the formal syntax, a global constant can only have a pointer type synatically.
Now SBpass works for all olden-ccured cases that we should support.
Times:
./bisort.exe 5000000 0;    7.8s
./em3d.exe 30000 300 50;   47.8s
./health.exe 8 250 1;      14.2s
./mst.exe 4000;            7.8s
./treeadd.exe 24 10;       5.6s
This is much slower (2-4x)than SoftBound. And the instrumented code is not
optimized before linking, because the LLVM optimizer gets failed at -O1 level,
the failure is from its Global Variable Opt pass..

The instrumentation of 129.compress is still failed, because the printer
of float value does not generate a right format that LLVM asm reader can
understand.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2956 eef5ea85-cc2e-0410-bd84-0fae1f012da8