Resigned inttoptr and ptrtoint. We should distinguish two kinds of ptr: at global spaces, and at heap or stack.
The first kind of ptr has a known address at compile time, and at
runtime their addresses cannot be reused; the second kind of ptr has no such
properties.

So, we can support i2p and p2i for the first ptr w/o parameterizing Mem
everywhere (at const2GV and getOperandValue), because we can maintain a
fixed mapping that is created at program initialization.

For the first kind of ptr, p2i is total. i2p can be undef if the int
value is not in the map.

This makes values in registers hold the substitution properities. If
const2GV is with Mem, that means its result can be affected by memory state, so
we cannot substitute it arbitrarily.

If const2GV is parameterized by Mem, and we do not distinguish the two
kinds of ptr, that complicates proofs, because we need to argue that
  1) memory model does not reuse addresses for globals, this is true for
     our corrent memory model, because it has inifite memory, and never
     reuses, but needs more work if we support finite memory later.
 2) It is hard to define simulation relations between the pointers or
    intergers casted from two programs, because related pointers can be in
    different memory addresses, casting them int can generate int values that
    lead to different computation states (if the value is used at
    branch).

This also indicates that the 2nd kind of ptr should eval to (the LLVM
undef) by i2p or p2i --- a set of values, because what their values are depends 
on runtime and platform. 

This change simplies the SoftBound proofs a lot, because the property
that constant values do not depend on Memory is true by default, needs
no reasoning.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3011 eef5ea85-cc2e-0410-bd84-0fae1f012da8