1) fixed all the admits about simulation relations. Our extended memory simulation relations are not bijective mapping between source and target programs, the interesting case is that undef values are mapped to anything, (other kinds of values are mapped exactly w.r.t a mapping function), so this means if a source program uses an undefined value as a conditional value for jmps, the target program can jmp to other locations.
The root problem is that our semantics is determinastic, and takes a
fixed value for undef, say 0. The non-det version of the semantics
should not have the issue, since no matter where the target program
jumps to, it is a valid state included by its source program.

With the det-opsem, we proved that if undefined values are not used for
jmp conditions, size of memory allocations, indices of GEP, the target program 
preserves its original program's state.

2) The rest admitted lemmas are all about freshness --- SBpass correctly
inserts fresh variables.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3021 eef5ea85-cc2e-0410-bd84-0fae1f012da8