1) changed IN[entry] to be nil in dominator analysis. it was [entry]. This did not affect the dom_entry and dom_successors property, but isn't correct, because IN[B] should be the blocks that strictly dominates B.
2) proved the correctness of dominator analysis w.r.t a strong
condition:
  B is strongly reachable if all its predecessors are strongly
reachable.
  If B is strongly reachable, then the result of its dominator analysis
is indeed what we wanted --- every CFG path to B must go through the
dominators.
  The condition is needed because the initial value of the analysis is
inconsistent --- every non-entry block is dominated by all blocks.
Strongly reachable block can be refined to be more consistent. But, we
may be able to prove it with a weaker condition, a result of a node
is consistent as long as one of its predecessor is reachable.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3279 eef5ea85-cc2e-0410-bd84-0fae1f012da8