Proved the correctness of dominator analysis.   1) the strong reachable condition used in the last revision is wrong, CFGs with loops do not hold the condition. Actually, we proved that the analysis is sound for a block if there is a directed walk (repetition of vertexes and arcs are fine) from the entry to the block.
  2) we also proved reachable blocks hold
     a) transitivity b) acyclic c) all blocks that strictly dominate a
block are totally ordered w.r.t strict domination relation

     Here, we resigned a complete/sound reachable checker that can be of
either Set or Prop type in Coq, so extractable. Note that the
reachablity analysis based on iteration recursion does not work because
it may fail when the iteration step is used up, at that case,
conservative answer (reachable) does not make sense, since the
correctness of dominator analysis is sound only when a block is
reachable precisely.
     
     The current reachable checker is based on the GraphBasics lib,
inductively defined, with complexity O(n*n), since at each inductive
step, we check if the added vertex is fresh, and if the added edge is
for existing vertexes.

     (c) is the core property for constructing dominator trees, which is
not implemented yet.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3280 eef5ea85-cc2e-0410-bd84-0fae1f012da8