Defined a coinductive divergence based on well-formedness, and proved that it equals to the other two versions of divergence. This finished the top-level simulation proofs of phinode-placement.
NOTE: In a coinductive constructor, Coq does not allow to use an inductive 
proof that uses the coinductive hypothesis because of its syntax
restriction. For such proofs, we should swap the inductive proofs and
and the coinductive hypothesis to let coinductive constructors see the
nested coinductive hyp directly, which is checked by Coq.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3782 eef5ea85-cc2e-0410-bd84-0fae1f012da8