1) fixed GEP to set the dynamic type of its return correctly. The old version always set it to be i1*.
2) fixed extraction and Ocaml compilation. This part could be broken >=
r3789 when we merged the Makefile infrastructure from src for 2.6. The
problem is that because of the extraction bug of Coq, we must use
fixextract.py to fix Coq output. However, Makefile ignores the Ocaml
compilation of these extracted files, so, we did not find fixextract.py
was not used.

Between r3879 and this commit, extracted interpreter and mem2reg may not
work. It was fixed. To make the version from r3789 work, refer the changes in
src3.0/Makefile and src3.0/Extraction/extraction_core.v.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3805 eef5ea85-cc2e-0410-bd84-0fae1f012da8