1) The memory simulation of SAS was not low enough to specify invariant precisely. It defined simulations in term of mstore and mload that actually only tell us the properties of codom of the low-level Mem.load/store, but not exactly the status of data in Memory.
For example, given
  M1 ~ M2 == forall ptr, mload M1 ptr = mload M2 ptr
  mstore M1 ptr0 gv = Some M1'
  mstore M2 ptr0 gv = Some M2',
we cannot show M1' ~ M2' == forall ptr, mload M1' ptr = mload M2' ptr

Consider at (blk,1) in M1 there are: undef; undef
         at (blk,1) in M2 there are: undef; 0:i16
we have 
  mload M1 ((blk,1),i16) = mload M2 ((blk,1),i16) = undef;undef
Suppose we do 
  mstore M1/M2 (blk,0) (0,i16) = Some M1'/M2'
then 
  mload M1' ((blk,1),i16) = 0:i16
  mload M2' ((blk,1),i16) = undef;undef
FAIL.

To solve the problem, we can set up a low-level memory relation for blocks and 
offsets, which is defined in sas_msim.v:
  a) between the SAS pair, suppose the promotable alloca with t is at blk0.ofs0,
     the simulation equates all contents at any valid blk and ofs except 
     [ofs0, ofs0+size t) at blk0
  b) otherwise, the two memory states equal.

To conclude all proofs, we need to show one more invariant about
promotable allocas: suppose the alloca of type t is blk.ofs, in blk, the only
valid memory range is [ofs, ofs+size t). With the invariant, when a
program meets the second store in the SAS pair, the store only and
exactly updates data in [ofs, ofs+size t) at blk. Such that, although
the simulation does not keep track of the data in the range, after the
store, they must match.

2) merged some memory simulation properties in SoftBound/Transform in to
Vellvm, so they can be shared.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3826 eef5ea85-cc2e-0410-bd84-0fae1f012da8