* fixed admitted cases in the top simulation lemmas of LAA/LAS and some properties of LASInfo/LAAInfo:
The old version proofs of id_rhs_val were correct, but not sufficient to
LAS. It assumed (vev_defs and substable_value) that 
  in f, if v2/v1 are only defined by cmds (non-phinodes/calls/arguments), 
  then [|v1|]=[|v2|].
However, in LAS, the value stored into the first store can be defined by
phinodes, calls or arguments.

So, we reproved id_rhs_val under a weaker condition:
  in f, if v1 >> v2, v2 is only defined by cmds, then [|v1|]=[|v2|].
In LAS, v2 is defined by load, v1 can be defined by anything, and we
know v1 >> v2.

We should also be able to extend id_rhs_val to allow v2 to be phinodes.
To allow v2 to be calls, we have to know the call does not use any
locations that the instructions between v1 and v2 use. 
It is not useful to allow v2 to be an argument for intraprocedural 
optimizations, since they do not change arguments.

I cannot see how the old id_rhs_val can be used by other SSA-based
optimizations. We can 'roll-back' to that version if necessary.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3844 eef5ea85-cc2e-0410-bd84-0fae1f012da8