** (palloca_props.wf_State) proved that when a program reaches a promotable alloca that defines pid, [|pid|] is None.
This property is important to prove SAS/DSE: DSE memory simulations
do not consider promotable locations; SAS mem-simulations ignore
promotoable locations between a SAS pair. To set up the simulation, we
need to know which location can be ignored.

In mem2reg, we only promote allocas in the entry block, such allocas
only run once. So we can look up the current stack frames, find all
defined variables by promotable allocas, whose values are the ignored
locations. So, palloca_props.wf_State is sufficient for mem2reg.

In general, a dead-store-elimination pass can work for arbitrary heap or
stack allocations---that can be in non-entry blocks, say in a loop. In
that case, looking up the current program state cannot find all ignored
locations. There are at least 3 solutions to solve the problem:

Suppose a store to id0 is dead, 

1) we consider a set of ignored locations IGN w.r.t id0---any value in the
set is a value that the program assigned to id0. The invariant is 
  if the current state is S, for any v in IGN, 
  then there exists a S', S' reaches S /\ S' defines id0 to be v

2) define a variant of operational semantics whose local state is a map
  id -> option (list value)
  Here, one id can have values with different versions. The runtime only
uses the latest one, while proofs or analysis use versions to keep track
old values that have been overwritten.

We should prove this semantics implies the invariant like 1).
Say id0 maps to v0, v1, ..., vn-1, vn at prog state Sn,
then there exist S0 S1, ..., Sn-1 that construct an execution sequence
to Sn.

The semantics formalizes one difference between SSA and lambda-calculus.
In SSA, single-assignment is only static, while in lambda-calculus,
single-assignment is both static and dynamic by substituting code. The
version-like semantics actually records the effects of
variable-overwriting in SSA, and simulates the code copy in
lambda-calculus: In SSA, a variable can be redefined at runtime iff the
variable is in a loop; if we consider a loop as a local function, each
new version the semantics created for the variable is like the
substitution that the standard application rule does.

3) The problem is also caused by not distinguishing readable and
unreadable locations when defining mem-relations. Concretely, the
current memory relatioins consider all memory locations except ignored
ones, although some of them are not readable by any pointers. For
example, in the case, the old version of id0 cannot be loaded by other 
pointers. 

Therefore, if a memory simulation only condisers locations that can be
loaded by some accessible variables in the program, the problem might also 
be solved??

** After solving this problem, we found that the memory simulation of
DSE has a similar problem like the problem of SAS that we had already
solved (See r3826), which is the next mission.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3852 eef5ea85-cc2e-0410-bd84-0fae1f012da8