1) reproved DSE with the low-level memory simulation that SAS uses.
The proof structures of DSE and SAS are very similar, should be able to
merge into a more general proof.

2) proved a new property for promotable locations: if a block b is
promotable for a alloca with type t, the memory range of b must contain
[0, size of t). This property was used by DSE. 

Consider the program state that returns from a call, and inside the call there 
is a promotable location at b. In the call, the simulation ignores
b.[0, size of t). When the call is returned, the simulation does not ignore 
b.[0, size of t) because the promotable alloca does not exist in the current 
stack frames. This is fine because that alloca at b was already deallocated, 
and never accessible. With the proved property, we know that the
allocated range of b contains the memory range the simulation ignored.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3857 eef5ea85-cc2e-0410-bd84-0fae1f012da8