applied the dominator lemmas to mem2reg. Found 3 problems to fix: 1) the wf_operand and wf_phi_operands do not allow using arguments! This is a bug of typing rules.
2) the current metatheory of dominator-relations does not support
commands that do not define variables, such as store, call that returns
void. This is not a problem for GVN, which only optimizes commands that
define variables. But, mem2reg and applications that reply on memory
operations also need dominator-relations of store. 

3) the current metatheory does not support dominating variables defined
by phinodes. Again, the current GVN does not value phi-nodes, while
mem2reg needs to deal with phinodes.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3863 eef5ea85-cc2e-0410-bd84-0fae1f012da8