fixed the 3rd problem in r3863: in the dominator relation a >> b, b can be phinodes.
The proofs w/o this were not wrong. In the existent proofs, we only
considered dominated commands and terminators. The well-formedness of
a phinode say that its incoming values must be dominated by the end of
the correspoinding incoming blocks. So, we never proved properties about
dominated phiniodes.

In mem2reg, we need to prove more lemmas about transitivity of
domination. In a>>b, a can be phinodes, so we need to consider dominated
phinodes.

All existent metatheory were extended for the change. We also proved
the acyclicness of idDominates and cmds_dominates_cmd that are needed by
id_rhs_val.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3866 eef5ea85-cc2e-0410-bd84-0fae1f012da8