Changed the mutual induction principle in typings.ott.
Using "Induction" instead of "Minimality" in Scheme for inductive
Prop's makes the induction principle weird, because the generated
principle can be used to prove things about propositions that depend
on the wf_* predicate itself. That leads to us proving statements of
the form

    Definition wf_something_prop
        a b c (H : wf_something a b c) := ...

    Theorem t : (forall a b c H, wf_something_prop a b c H) /\
        ...

instead of the more natural

    Definition wf_something_prop
        a b c := wf_something a b c ->

    Theorem t : (forall a b c, wf_something_prop a b c) /\
        ...

While trying to remove mutual induction from the development, it was
really hard to rewrite the theorems that had this form. Hopefully this
should make the task a tad easier now.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3871 eef5ea85-cc2e-0410-bd84-0fae1f012da8