Converted unnecessary mutually inductive definitions.
Notice that this requires an upgrade to Ott 0.21.2.

In order to change the original development as little as possible, we
tried to create backward-compatible versions of the most important
mutually inductive principles. In particular, lemmas such as
wf_value_mutind needed to be restated in terms of a new version of
wf_value_list, and similarly for other inductively defined predicates.

Wherever wf_value_list was used in premises, Ott now generates
different kinds of premises. Of course this is a big problem when
using old lemmas that were stated in terms of wf_value_list. A few new
tactics were created in typings.v to cope with these problems. For
instance, find_wf_value_list will try to find a hypothesis in the new
format in the context and convert it to wf_value_list so it can be
used with the old lemmas.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3899 eef5ea85-cc2e-0410-bd84-0fae1f012da8