1) refractored the CFG-presevation proofs (block-domination, reachability, succ, pred, entry, ...) in Vminus to the module in trans.v. subst.v/vsubst.v/motion.v were changed to use the module for shared code. insert.v and removal.v should use the module, too. We may change them later. All passes in mem2reg.v can also use the module.
The invariant of the transformations is that they do not change CFG. So,
the CFG-presevation proofs are the same.

There is an inconvenient issue when using the module. A transformation
is parameterized as a structure of the module, like

Struct pass := mkPass {
ftrans: fdef -> fdef
}

When using the module with a pass, say 
  subst: fdef -> fdef

We may want to instantiate the ftrans field by subst. But, then the
proof contexts may contain pass.(ftrans) and subst together, and Coq
cannot automatically convert the ones in hypothesis, but can convert the
ones in goals. This makes proofs seem not generic, since we have to
explicitly 'replace' one by the other sometimes. And idea?

2) Vminus is not compilable by ott 0.21.2 for unknown reasons, 
0.20.1 still works.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3945 eef5ea85-cc2e-0410-bd84-0fae1f012da8