implemented and verified the dead-phinode-elimination pass to make vmem2reg's final output close to the pruned SSA that LLVM's mem2reg generates.
1) 
This pass uses die.v directly. So, it does not remove aggressive dead
phinode. One such an example is in the "compress" function of
129.compress, where there are two phinodes that use each other, but are
not used by others:

  l1:
    x = phi [y,l2] [0,entry] ...

  l2:
    y = phi [x,l1] [0,entry] ...

Here is another aggressive case in the "oops" function of 129.compress:

entry:
  %p = alloca 

  br .. l1 l2

l1: 
  
  br l2

l2:
  call @exit() 
  unreachable

dead:
  %r = load %p
  ret %r
 
The number of such cases is very small in the benchmarks.

2) We added a new flag -dead-phi-elim to vmem2reg. Experimentally, I could not 
see the performance is changed by deleting the dead phinods, although a lot of 
dead phinodes were removed. So, by default the flag is false. 



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@3983 eef5ea85-cc2e-0410-bd84-0fae1f012da8